Shailesh changes

+++ cms/models.py
@@ -31,7 +31,7 @@ class College(models.Model):
     email = models.EmailField()
     description = RichTextField()
     gallery = models.ManyToManyField(Gallery)
-    rating = models.PositiveIntegerField()
+    rating = models.DecimalField(max_digits=2, decimal_places=1)
     date_created = models.DateTimeField(auto_now_add=True)
     author = models.ForeignKey(Author, on_delete=models.SET_NULL, null=True, blank=True)
